refactor(review) :메서드 명 isNoop ->  isNewUser로 변경 -dto 패키지 변경
