+++ src/main/java/com/delivery/justonebite/review/application/dto/request/UpdateReviewRequest.java
@@ -1,4 +1,4 @@
-package com.delivery.justonebite.review.presentation.dto.request;
+package com.delivery.justonebite.review.application.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.Max;
