+++ src/main/java/com/delivery/justonebite/shop/presentation/dto/response/ShopReviewResponse.java
@@ -1,6 +1,6 @@
 package com.delivery.justonebite.shop.presentation.dto.response;
 
-import com.delivery.justonebite.review.presentation.dto.response.ReviewResponse;
+import com.delivery.justonebite.review.application.dto.response.ReviewResponse;
 import io.swagger.v3.oas.annotations.media.ArraySchema;
 import io.swagger.v3.oas.annotations.media.Schema;
