+++ src/test/java/com/delivery/justonebite/review/application/service/ReviewServiceTest.java
@@ -8,10 +8,10 @@ import com.delivery.justonebite.order.domain.enums.OrderStatus;
 import com.delivery.justonebite.order.domain.repository.OrderHistoryRepository;
 import com.delivery.justonebite.order.domain.repository.OrderRepository;
 import com.delivery.justonebite.review.entity.Review;
-import com.delivery.justonebite.review.presentation.dto.request.CreateReviewRequest;
-import com.delivery.justonebite.review.presentation.dto.request.UpdateReviewRequest;
-import com.delivery.justonebite.review.presentation.dto.response.CreateReviewResponse;
-import com.delivery.justonebite.review.presentation.dto.response.ReviewResponse;
+import com.delivery.justonebite.review.application.dto.request.CreateReviewRequest;
+import com.delivery.justonebite.review.application.dto.request.UpdateReviewRequest;
+import com.delivery.justonebite.review.application.dto.response.CreateReviewResponse;
+import com.delivery.justonebite.review.application.dto.response.ReviewResponse;
 import com.delivery.justonebite.review.repository.ReviewRepository;
 import com.delivery.justonebite.shop.domain.entity.Shop;
 import com.delivery.justonebite.user.domain.entity.User;
